problem about <img>
The default display value of the <img> tag is inline or inline-block.
When images are mixed with text, there may be some spacing that looks like margin but is actually due to the layout characteristics of the inline elements,
such as the white space at the bottom of the image related to the line height. In this case display: block;